E-Vahan an android based application.
In today’s world of internet everything is done online to minimize human
effort and to maximize productivity. In the big cities alone most of the
time is wasted on finding a speedy public transport vehicle like auto
rickshaw, taxi etc. This problem needs to be solved so that the daily
commuters can find an effective means of transport as soon as possible
so as to reach their destination quickly.
So what we propose here is to develop a commercial android application
which helps people in find a transport vehicle in the near vicinity of
their current location.
The app can work both offline and online mode depending upon the
availability of reliable data connection like Wi-Fi or internet. The
user can find the contact numbers of any auto rickshaw or taxi drivers
or any transport agency in the vicinity of their location.